+++ examples/multiFileProject/multiFileProject.cpp
@@ -1,0 +1,12 @@
+/****************************************************************************************************************************
+  multiFileProject.cpp
+  AsyncUDP_WT32_ETH01 is a Async UDP library for the WT32_ETH01 (ESP32 + LAN8720)
+
+  Based on and modified from ESPAsyncUDP Library (https://github.com/me-no-dev/ESPAsyncUDP)
+  Built by Khoi Hoang https://github.com/khoih-prog/AsyncUDP_WT32_ETH01
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#include "multiFileProject.h"