+++ examples/multiFileProject/multiFileProject.h
@@ -1,0 +1,17 @@
+/****************************************************************************************************************************
+  multiFileProject.h
+  AsyncUDP_WT32_ETH01 is a Async UDP library for the WT32_ETH01 (ESP32 + LAN8720)
+
+  Based on and modified from ESPAsyncUDP Library (https://github.com/me-no-dev/ESPAsyncUDP)
+  Built by Khoi Hoang https://github.com/khoih-prog/AsyncUDP_WT32_ETH01
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#pragma once
+
+#define _ASYNC_UDP_WT32_ETH01_LOGLEVEL_       1
+
+// Can be included as many times as necessary, without `Multiple Definitions` Linker Error
+#include "AsyncUDP_WT32_ETH01.hpp"